- Enabled SharedPreferences so that users can select their current store (list of stores is pulled down automatically from the Parse server)   - Store selector is accessed through in Settings   - SelectedProduct view shows the stock for the current store - Reserving stock in a store now successfully makes changes to a database - Removed old unused XML files

+++ app/src/main/java/com/example/InfiniTeam/models/ProductModel.java
@@ -9,11 +9,11 @@ public class ProductModel  implements Serializable {
     private String description;
     private String extra1;
     private String extra2;
-    private int price;
+    private float price;
     private int stock;
     private String image;
 
-    public ProductModel(String id, String name, String description, String extra1, String extra2, int price, int stock, String image) {
+    public ProductModel(String id, String name, String description, String extra1, String extra2, float price, int stock, String image) {
         this.id = id;
         this.name = name;
         this.description = description;
